Anytime Anywhere Electricity Bill Generator
Inspiration

The year 2020 was a bizarre situation for the world and there was a major setback that happened in all fields we started to live indoors, many regular jobs went remote, financial instability is a few of the many challenges that we have faced. As social gathering and working in a closed atmosphere started to create an adverse effect on the spreading of the pandemic, job nature also changed. Companies started to establish work from home strategy as a part of a business continuity plan in both the private and government sector. A commonality in all this is the lack of mobility for people to complete their day-to-day activities. This created a serious impact in sectors and processes where human or manual interventions are required, especially in the sector of electricity Bill generation. 
Electricity bills started to show a considerable amount of variation due to the absence of manual works and the bill was generated based on the previous billing cycles, which is not precise. Domestic, commercial, and industrial sectors started to pay huge amounts of bills with no proportion to the number of units consumed, so we decided to come up with the finest solution for this enigma. 

 

 




Why do we need

Majorly there are three categories of meters for measuring electricity, Electromechanical induction type energy meter, Electronic Energy Meters, and Smart Energy Meters. Talking about Electronic Energy Meters, there are millions of meters which are functioning all over India. These electronic energy meters show the number of units consumed and the manual calculation of charges will be done accordingly. An individual from the government sector specifically from the electricity board needs to visit every property and record the units which are later entered into the portal for calculation. This is always a tedious task to do and there are many setbacks for this manual process such as tiring travels to the isolated location, wrong bill amount due to incorrect meter reading, and risk of spreading contagious diseases such as COVID-19, Dengue, Malaria, etc. Here we come up with the innovative thought using the tool Automation Anywhere who pioneered RPA solutions. All we have to do here is Snap the image of the energy meter measurement, Send it to the email id, and Pay your electricity bills. All these manual efforts can be replaced in just a click-and-send button and we call it ‘SNAP IT SEND IT Pay IT’. 
For the Prototype study, we took Tamil Electricity Board. The total area covered is around 130,058 square kilometers and the population of the area covered by TNEB is 624 Lakhs approx. TNEB caters to a total of 200 Lakhs consumers and has a total of around 70000 employees currently. Total revenue collection for the year 2020 – 21 was Rs, 2,24,739 Crores, and revenue collection efficiency was 98 %(values and figures are subject to change based on the current survey). As per the recent reports, India is losing 100,000 crores in unbilled electricity, and the End- End solution of electricity bill generation, we bring up to overcome this issue will resolve all this huge loss. Also, it saves a lot of time rather than visiting the places for noting the units later the calculation and entry to portal, considering the number of employees and manual effort behind the bill calculation we can reduce the FTE used for the process, its cost-saving for both the customer and sector and more-over there will be a work life balance for whom does that all manual work.

How we built it
 
The End- End Solution for EB bill generation is built on A360 and also integrates with various other technologies and application.
•	IQ bot
•	Python
•	Google vision OCR
•	Machine Learning
•	Microsoft Access
•	TNEB Bill calculator(https://www.tnebnet.org/awp/tariffMaster;jsessionid=j2w0O5CkqEarl3ff1vCe+o8Y?execution=e1s1)

The entire calculation of the bills is based on the Image send out by the user of the units consumed. This image is sent out to a particular email Id which is the first and foremost thing the consumer needs to do. When the bot is triggered, it reads the email and pushes the image of meter reading to IQ bot as input (Bot only accepts email from registered email ID). IQ bots do scrapping of units using the Google vision as OCR and the bill is calculated on the TNEB website.

For predictive analysis of the electricity bill, we used simple linear regression for modeling the relationship between each month (independent variable) and units consumed (dependent variable). This Machine Learning model will give a close predicted output to the observed values given. But how does this helps us? The answer is, it will check whether the calculated units based on the meter reading and the predicted output don’t have a greater difference. We have set a threshold value of 20 percent as the deviation whether positively or negatively. This will help to capture any fraudulent images send out or any errors in calculation. When we observe a greater deviation (which we set as 20%) then discrepancies will be reported to the administrator in the amount and in that case a person from the electricity board can visit the consumer to the credibility of the meter reading.